Adding Merge functions for enumerators.
This is not an inverse function to Tee. Tee duplicates entries in a channel and feeds them for separate processing. Merge aggregates values from multiple channels. It is differenitiated from a Union function in that it does not attempt to remove duplicates.

The primary function of Merge is to allow for fan-in when processing of lists was being handled by multiple threads.