Adding a cancel channel to Enumerators
Basically, without having a means of canceling an Enumeration, Go
routines leak. If enumeration is stopped before the end is
reached, the Go Routines will never close. Because of that, the
channels will never get garbage collected.

Scenarios like "Any" are so important this this really has to be
done.